Layer: Add composition support
This also adds methods to create a non tilted projection matrix. This
requires backface culling to be set to none. UI::Manager gained a
compost method to update any layers that need to be recomposted. Bounds
gained a new method to transform (translate and scale) a bounds object
to a new bounds object.